fix: withdrawal amounts
